% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj_path.R
\name{adj_path}
\alias{adj_path}
\title{adj_path}
\usage{
adj_path(path)
}
\arguments{
\item{path}{character, the path to be adjusted}
}
\value{
The path without the ending forward slash.
}
\description{
Adjusts a path by removing, if present, an ending forward slash.
}
\examples{
\dontrun{
if(interactive()){
 adj_path('~/git/some_model/')
 }
}
}
