% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cumulative_returns.R
\name{calculate_cumulative_returns}
\alias{calculate_cumulative_returns}
\title{calculate_cumulative_returns}
\usage{
calculate_cumulative_returns(rets_xts)
}
\arguments{
\item{rets_xts}{xts, an xts of asset returns}
}
\value{
A matrix displaying the cumulative return for each column of returns.
}
\description{
Calculates the cumulative return for each column of an xts.
}
\details{
A minimal \code{PerformanceAnalytics::Return.cumulative()}.
}
\examples{
\dontrun{
if(interactive()){
 quantmod::getSymbols('SPY')
 SPY[,4] \%>\%
   {. / stats::lag(.) - 1} \%>\%
   stats::na.omit() \%>\%
   clhelpers::calculate_cumulative_returns()
 }
}
}
