% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_portfolio_return.R
\name{calculate_portfolio_return}
\alias{calculate_portfolio_return}
\title{calculate_portfolio_return}
\usage{
calculate_portfolio_return(asset_units_xts, verbose = FALSE)
}
\arguments{
\item{asset_units_xts}{xts, an xts of units to be joined to the asset returns to form a portfolio return}

\item{verbose}{boolean, passed to PerformanceAnalytics::Return.portfolio(), Default: FALSE}
}
\value{
See documentation for PerformanceAnalytics::Return.portfolio().
}
\description{
Adds columns for cash before calculating the portfolio return.
}
\details{
This is just a wrapper for PerformanceAnalytics::Return.portfolio().  See documentation for PerformanceAnalytics::Return.portfolio().
}
