% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_returns.R
\name{calculate_returns}
\alias{calculate_returns}
\title{calculate_returns}
\usage{
calculate_returns(closes_xts, ret_periodicity = "day")
}
\arguments{
\item{closes_xts}{xts, an xts of closing prices}

\item{ret_periodicity}{character, the return periodicity which is passed to xts::endpoints(), Default: 'day'}
}
\value{
An xts of asset returns.
}
\description{
Wrapper for TTR:ROC() but with easy periodicity changing.
}
\examples{
\dontrun{
if(interactive()){
 quantmod::getSymbols('SPY')
 SPY[,4] \%>\%
   clhelpers::calculate_returns('week')
 }
}
}
