% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_sd.R
\name{calculate_sd}
\alias{calculate_sd}
\title{calculate_sd}
\usage{
calculate_sd(rets_xts)
}
\arguments{
\item{rets_xts}{xts, an xts of returns}
}
\value{
A matrix of standard deviations.
}
\description{
A minimal PerformanceAnalytics::StdDev().
}
