% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_td_access.R
\name{get_td_access}
\alias{get_td_access}
\title{get_td_access}
\usage{
get_td_access(full_path_to_td_credentials)
}
\arguments{
\item{full_path_to_td_credentials}{character, the full path to the directory containing the consumer key and refresh token for the \emph{rameritrade} package}
}
\description{
Quick access to TD Ameritrade's API.
}
\details{
See the below under \emph{References} for directions on creating the relevant files.  Once they are created, put them in a directory and point this function to that directory for API access.
}
\examples{
\dontrun{
if(interactive()){
 get_td_access('~/td')
 }
}
}
\references{
\url{https://github.com/exploringfinance/rameritrade/}
}
