% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_col.R
\name{lag_col}
\alias{lag_col}
\title{lag_col}
\usage{
lag_col(x, col_num, k = 1)
}
\arguments{
\item{x}{matrix, xts, data frame or data table}

\item{col_num}{integer, the column number to lag}

\item{k}{integer, lag length, Default: 1}
}
\value{
Object of the same format with values in one column lagged.
}
\description{
Given a matrix, xts, data frame, or data table, lag the values of a specified column.
}
\examples{
\dontrun{
if(interactive()){
 library(magrittr)
 quantmod::getSymbols('SPY;DIA')
 x <- lapply(c('SPY','DIA'), function(x) get(x)[,4]) \%>\% do.call(what=cbind)
 x[1:5] \%>\% lag_col(1, 3)
 }
}
}
