% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_closes_xts.R
\name{make_closes_xts}
\alias{make_closes_xts}
\title{make_closes_xts}
\usage{
make_closes_xts(asset_symbols, n = 430)
}
\arguments{
\item{asset_symbols}{character, a string of asset_symbols}

\item{n}{integer, passed to utils::tail(), Default: 430}
}
\value{
An xts of closes.
}
\description{
You probably don't need to use this.
}
