% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dates.R
\name{make_dates}
\alias{make_dates}
\title{make_dates}
\usage{
make_dates()
}
\description{
Produces a list of dates on which the market is open; indicates start/end-of-month days, as well as end-of-week.
}
\details{
Creates a directory called \code{data} which contains the dates RDS file.
}
