% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_permutations.R
\name{make_permutations}
\alias{make_permutations}
\title{make_permutations}
\usage{
make_permutations(x)
}
\arguments{
\item{x, }{any kind of vector}
}
\value{
A matrix of all permutations.
}
\description{
Creates a matrix which shows all the permutations of a series.
}
\details{
Taken from Stack Overflow.
}
\references{
\url{https://stackoverflow.com/questions/11095992/generating-all-distinct-permutations-of-a-list-in-r?answertab=scoredesc#tab-top}
}
\author{
"Adrian"
}
