% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rets_xts.R
\name{make_rets_xts}
\alias{make_rets_xts}
\title{make_rets_xts}
\usage{
make_rets_xts(asset_symbols, n = 430)
}
\arguments{
\item{asset_symbols}{character, a string of asset_symbols}

\item{n}{integer, passed to utils::tail(), Default: 430}
}
\value{
An xts of asset returns.
}
\description{
Makes an xts of returns from a string of symbols.
}
