% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_wiki_dow_components.R
\name{scrape_wiki_dow_components}
\alias{scrape_wiki_dow_components}
\title{scrape_wiki_dow_components}
\usage{
scrape_wiki_dow_components(from_date = "2007-01-01")
}
\arguments{
\item{from_date}{character, representing a date in YYYY-MM-DD format, it indicates the date which fills in missing date values}
}
\value{
a data.table with two columns: 1) Symbol and 2) Date_Range (a range of days)
}
\description{
scrapes dates for when particular Dow components were added/removed in a timeframe
}
\examples{
\dontrun{
if(interactive()){
 clhelpers::scrape_wiki_dow_components()
 }
}
}
\references{
\href{https://en.wikipedia.org/wiki/Dow_Jones_Industrial_Average}{DJIA Wiki}
}
