% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_special_day.R
\name{show_special_day}
\alias{show_special_day}
\title{show_special_day}
\usage{
show_special_day()
}
\value{
A string indicating any of the three: EOM (end of month), SOM (start of month), EOW (end of week).
}
\description{
Shows if today is the first/last trading day of the week/month.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 clhelpers::show_special_day()
 }
}
}
